# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Workforce Management Platform
  version: 1.0.0
paths:
...
  /leave/snapshots/run:
    post:
      summary: ""
      description: ""
      tags: []
      responses:
        "200":
          description: Success
        "403":
          description: Forbidden
      security:
        - BearerAuth: []
  /leave/snapshots:
    get:
      summary: ""
      description: ""
      tags: []
      responses:
        "200":
          description: Success
        "403":
          description: Forbidden
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT